Enable client and endpoint for adding new attribute to FUW REST API
Bulk upload of metadata from spreadsheet requires that its controller
can add new attributes from the spreadsheet, and not just  replace them as it was done
from the attributes form in the javascript Vue component.

So we add the option to the ReplaceAction to not delete existing attributes, keyed on the requested route.

We now have two different routes to reach the ReplaceAction.php action, one from the attribute form in the wizard as before,
and a new one from spreadsheet upload workflow.

In addition to the change in the action, we add the corresponding client method in FileUploadService.php.

+++ fuw/app/frontend/actions/AttributeController/ReplaceAction.php
@@ -46,9 +46,18 @@ class ReplaceAction extends \yii\rest\Action
         // Yii::warning(var_export($attr,true));
         if( $attr && isset($attr["Attributes"]) ) {
 
-            // delete existing ones
-            foreach ($modelClass::findAll([ 'upload_id' => $upload_id ]) as $oldModel) {
-                $oldModel->delete();
+            if("attribute/replace" === Yii::$app->requestedRoute) {
+                // delete existing ones
+                Yii::info("Action mode: attribute/replace");
+                foreach ($modelClass::findAll([ 'upload_id' => $upload_id ]) as $oldModel) {
+                    $oldModel->delete();
+                }
+            }
+            elseif("attribute/add" === Yii::$app->requestedRoute) {
+                Yii::info("Action mode: attribute/add");
+            }
+            else {
+                throw new ServerErrorHttpException('Unrecognized action:'.Yii::$app->requestedRoute);
             }
 
             // create an array of models
